requirements.txt
asgiref==3.3.1
certifi==2022.6.15        
charset-normalizer==2.0.12
Django==2.1.15
django-cors-headers==2.2.0
django-pyodbc-azure-2019==2.1.0.0
idna==3.3
joblib==0.17.0
numpy==1.19.3
pybars3==0.9.7
PyMeta3==0.5.1
pyodbc==4.0.30
python-crfsuite==0.9.7
pytz==2020.4
pyvi==0.1
requests==2.28.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn-crfsuite==0.3.6
sqlparse==0.4.1
tabulate==0.8.7
threadpoolctl==2.1.0
tqdm==4.51.0
typing_extensions==4.2.0
urllib3==1.26.9